+++ R/AllGenerics.R
@@ -29,9 +29,6 @@ setGeneric("dropNaSpectraVariables", function(object, ...)
 setGeneric("export", function(object, ...)
     standardGeneric("export"))
 #' @rdname hidden_aliases
-setGeneric("filterIntensity", function(object, ...)
-    standardGeneric("filterIntensity"))
-#' @rdname hidden_aliases
 setGeneric("filterMzRange", function(object, ...)
     standardGeneric("filterMzRange"))
 #' @rdname hidden_aliases